Split Qwen35 prefill helpers
