+++ crates/runner/src/main.rs
@@ -17,6 +17,7 @@ mod prefill_engine;
 mod profiling;
 mod qwen35_dflash_engine;
 mod qwen35_kv_trace;
+mod qwen35_prefill;
 mod qwen35_runtime;
 mod qwen35_trace;
 mod qwen35_trace_utils;
